Refactor map helpers & extract market logistics
Rename map helper functions to get_* for clarity (area->get_area, metroarea->get_metroarea, direction_to->get_direction_to, arability->get_arability) and update all call sites in actions, military, and region modules. Extract market-related calculations and market-building logic from game/objs/market.py into a new game/logic/logistics.py (is_connected, get_production/consumption/supply/fulfillment, build_markets) to decouple the Market dataclass from state/logic. Update Region to use get_supply and get_arability and adjust region tile area handling to use get_area. Minor cleanups: remove unused TYPE_CHECKING/industry imports from market.py and drop unused dataclasses.field import in tile.py.
